+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## 0.17.0 (provisional)
 
 * Adding `HashedArrayTree`.
+* Adding `BitVector`.
 * Adding `#.frequency` to `MultiSet`.
 * Adding `#.grow` to `DynamicArray`.
 * Adding `#.reallocate` to `DynamicArray`.
 * Adding `#.resize` to `DynamicArray`.
 * Fixing several `MultiSet` issues.
+* Renaming `DynamicArray` to `Vector`.
 * Renaming the `DynamicArray.initialLength` option to `initialCapacity`.
 * Renaming `DynamicArray.allocated` to `capacity`.
 * Optimizing `MultiSet` performance.